+++ continuous_integration/travis/install.sh
@@ -74,7 +74,6 @@ fi
 pip install -q --upgrade --no-deps \
     cachey \
     graphviz \
-    pyarrow \
     pandas_datareader
 
 pip install -q --upgrade \
@@ -86,11 +85,27 @@ pip install -q --upgrade \
     xxhash
 
 if [[ ${UPSTREAM_DEV} ]]; then
+    echo "Installing PyArrow dev"
+    conda install -q -c twosigma \
+          arrow-cpp \
+          parquet-cpp \
+          pyarrow
+else
+    echo "Installing PyArrow"
+    conda install -c conda-forge \
+          arrow-cpp \
+          parquet-cpp \
+          pyarrow
+
+fi;
+
+if [[ ${UPSTREAM_DEV} ]]; then
     echo "Installing NumPy and Pandas dev"
     conda uninstall -y --force numpy pandas
     PRE_WHEELS="https://7933911d6844c6c53a7d-47bd50c35cd79bd838daf386af554a83.ssl.cf2.rackcdn.com"
     pip install -q --pre --no-deps --upgrade --timeout=60 -f $PRE_WHEELS numpy pandas
 fi;
+
 
 # Install dask
 pip install -q --no-deps -e .[complete]